requirements.txt
amqp==5.3.1
asgiref==3.9.1
asttokens==3.0.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.1
decorator==5.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
executing==2.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipdb==0.13.13
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kavenegar==1.1.2
kombu==5.5.4
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
